Prevent prevent contents explosion proc (#47337)
About The Pull Request

Proc literally only existed to return a boolean. Contents explosion is a clusterfuck anyhows. Basically it works something like this:

If the turf is the epicenter: GetAllContents of the contents and call ex_act on them directly. But it doesn't get the contents of whatever returns true on prevents_contents_explosion.
Otherwise: Call ex_act on the turf which calls the contents_explosion proc which has nothing to do with the prevents_content_explosion proc.

This is barely the smallest droplets of water onto a hot stone.

PR replaces the proc with a bitflag in flags_1.