[READY] (nerf) Adds proximity crew pinpointer to Paramedics (#48418)
About The Pull Request

When #48236 was in progress, there was discussion about how a crew pinpointer was OP for paramedics. I immediately had an idea for a new one and made a PR to the paramedic PR. 81Denton asked me to make this its own PR after the paramedic PR was merged.

The proximity crew pinpointer does not point in the direction of the patient, but has a circle sprite that gets smaller larger as you get closer, so you have to search around, or combine it with the handheld crew monitor.

The circle is at maximum the smallest size after 29 tiles away, so you must be 28 or fewer tiles away for an indication that you are getting close. I welcome feedback on this. The handheld crew monitor can be used to find the area they are in.

proxpin

Demo video: https://youtu.be/_rXeJ49b-e8 (The pinpointer gets bigger rather than smaller now)

This is only attainable at the moment to paramedics as they spawn. Should it also be available elsewhere? now constructable from the Biological Technology techweb
Further changes

Moved Crew Pinpointer into premium nanomed items for 900 credits
Moved Handheld Crew monitor into premium nanomed items for 1750 credits, same price as before
Why It's Good For The Game

In #48236 it was stated that paramedics should start with the bare minimum equipment to do their job. The crew pinpointer that points them in exactly the right direction is pretty powerful, and this nerf will encourage them to either get used to the proximity one, or upgrade to the normal one at a price from the NanoMed or from techwebs.

Moving the Crew Monitor into premium was necessary to make it more of a luxury than an instant grab.
Feedback for nanomed premium prices on the normal pinpointer and the handheld monitor would help a lot
Changelog

cl JJRcop
add: New proximity crew pinpointer.
balance: Paramedic's crew pinpointer replaced with proximity crew pinpointer.
balance: The crew monitor was moved to premium items in nanomed, same with the handheld crew pinpointer
/cl