Enabling random lawsets for AI at round start
Had a bit of a discussion and wanted to make the AI lawsets at round start randomised. The files themselves may not change but this is mostly to spark some discussion as the configs themselves will be changed on the servers end rather than on the github mostly. So please gives some more suggestions what lawsets should be nabled like asimovpp, crewsimov, and all that.

The lawsets that I have active are:
Asimov
Paladin5
Corporate
Drone
Reporter