[Port] Rats Rats we are the rats! (#2144)
* Rats take 2

* Does this satisfy you, travis?

* 1 in a million

* Whats wrong travis? Dont like my indents?

* The rats can dig up some nice items now.

* BAD  AT M A T H

* Update regalrat.dm

* Sentient regal rats can now add things to their faction

* Hostile rats now have a message telling you they die.

Its better than attacking a rat and having it de-atomize without any kind of information

* People who observe the enlist ability will see a notice now, same with the person who got added to the faction

* Changed the chance to roll 'meme loot"

Changed it from one in a mil to one in hundred thousand. You roll a 7 sided dice, 2 sides are actually useful (pulse rifle or anti bread) and the other 5 sides are useless. Also pulse rifle and anti bread are admin logged when they spawn.

* Maintspawn with royal cheese and a mouse

* whoops