Port harddel fixes, as well as "Fixes stasis bed runtimes" (#4598)
* Fixes a bunch of harddels that are sourced from player action (#59371)

Sourced from #59118 and a cursed project I'll pr later, This pr contains a lot of harddel fixes for stuff that pops up after a player interacts with something. I'm not gonna list them all here because there's something like 60 130, check the commit log if you're curious

Oh and I moved ref tracking screaming to a separate define, and made some optimizations to the thing in general. I think that's it, this pr is a bit of a frankenstine

* Activate Held Object and Drop hotkeys (Z and Q by default) now activate and store arm implant tools (#53893)

* Self_attack for arms if the hand is empty

* Oops

* Light runtime fix

This is not my runtime, but I will fix it all the same

* return better good

* How about a variable name that actually fits

* asdf

* Fix some bad merges

* Fixes stasis bed runtimes (#50525)

* fix

* suggested changes

* Theresnotwoargument

* changes

* okay

* Update code/game/objects/structures/tables_racks.dm

Co-Authored-By: Rohesie <rohesie@gmail.com>

Co-authored-by: Rohesie <rohesie@gmail.com>

* Fix runtime when emitters shoot SM. I don't understand why this started happening, but the runtime is from trajectory being null.

* Fix ctf flag var type

* Revert "Activate Held Object and Drop hotkeys (Z and Q by default) now activate and store arm implant tools (#53893)"

The changes will be made part of a separate PR

* Fix runtime due to bad augment rework revert

* Replace a couple QDELETED with QDELING

Co-authored-by: LemonInTheDark <58055496+LemonInTheDark@users.noreply.github.com>
Co-authored-by: zxaber <37497534+zxaber@users.noreply.github.com>
Co-authored-by: Archanial <37456678+Archanial@users.noreply.github.com>
Co-authored-by: Rohesie <rohesie@gmail.com>